+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Added
+- `SButtonGroup` and `SButtonGroupItem` tests.
+- `SButton` tests.
 - `STextArea` tests.
 - `SLabel` tests.
 - `SBreadcrumbs` and `SBreadcrumbsItem` tests.